AppVeyor: fully automated build of 32- and 64-bit Windows binaries
For pull requests and pushes to master just build 32-bit binaries
and do not store artifacts.

Whenever a tag is pushed to the "release" branch - produce
full set of 32-bit and 64-bit binaries, create a GitHub
release (if does not exist already) and attach the binaries
to the release.

Issues open:

http://bit.ly/twoBuilds

 AppVeyor starts two builds in response to the "new commit"
 and "new tag" events.

http://bit.ly/bad_build_success

 Parse error in the appveyor.yml causes the system to
 silently ignore the commit.

http://bit.ly/appvLessDup

 Is there a way to remove duplicate stanzas from
 both configurations?